fix: add some comments and aggregate some utils
