+++ bcs-k8s/bcs-federated-apiserver/pkg/configuration/cluster.go
@@ -16,56 +16,39 @@ package configuration
 import (
 	"context"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
+	"github.com/Tencent/bk-bcs/bcs-k8s/bcs-federated-apiserver/pkg/utils"
+
 	"k8s.io/klog"
 	fedv1b1 "sigs.k8s.io/kubefed/pkg/apis/core/v1beta1"
-	"sigs.k8s.io/kubefed/pkg/client/generic"
 )
 
+// AggregationClusterInfo is the struct of final member cluster list.
 type AggregationClusterInfo struct {
 	memberClusterList string
 }
 
-//var ClusterInfo AggregationClusterInfo
-
+// SetClusterInfo is the function for setting the member clusterList info.
+// if the configmap has set the "memberClusterOverride", using it;
+// otherwise using the clusterList from the KubeFedCluster resource of the namespace kube-federation-system.
 func (aci *AggregationClusterInfo) SetClusterInfo(acm *AggregationConfigMapInfo) {
-
 	if acm.GetClusterOverride() != "" {
 		klog.Infoln("Get memberClusterList from AggregationConfigMapInfo of ClusterOverride.")
 		aci.memberClusterList = strings.ToUpper(acm.GetClusterOverride())
 	} else {
 		klog.Infoln("Get memberClusterList from kubeFederated member cluster.")
 
-		config, err := rest.InClusterConfig()
+		clientSet, err := utils.NewKubeFedClientSet()
 		if err != nil {
-			// fallback to kubeconfig
-			kubeconfig := filepath.Join("~", ".kube", "config")
-			if envvar := os.Getenv("KUBECONFIG"); len(envvar) > 0 {
-				kubeconfig = envvar
-			}
-			config, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
-			if err != nil {
-				klog.Errorf("The kubeconfig cannot be loaded: %v\n", err)
-				os.Exit(1)
-			}
-		}
-
-		clientSet, err := generic.New(config)
-		if err != nil {
-			klog.Errorf("Failed to create clientset: %v", err)
+			klog.Errorf("Failed to create clientSet: %v", err)
 			os.Exit(1)
 		}
 
 		clusterList := &fedv1b1.KubeFedClusterList{}
 
 		for {
-			aci.memberClusterList = ""
-
 			err = clientSet.List(context.TODO(), clusterList, "kube-federation-system")
 			if err != nil {
 				klog.Warningf("Error retrieving list of federated clusters: %v\n", err)
@@ -73,17 +56,8 @@ func (aci *AggregationClusterInfo) SetClusterInfo(acm *AggregationConfigMapInfo)
 				if len(clusterList.Items) == 0 {
 					klog.Errorln("No federated clusters found, wait for join KubeFed member cluster")
 				} else {
-					for _, cluster := range clusterList.Items {
-						var clusterTmp string
-						if acm.GetClusterIgnorePrefix() != "" {
-							clusterTmp = strings.TrimPrefix(cluster.Name,
-								acm.GetClusterIgnorePrefix())
-						} else {
-							clusterTmp = cluster.Name
-						}
-						aci.memberClusterList += strings.ToUpper(clusterTmp) + ","
-					}
-					aci.memberClusterList = strings.TrimRight(aci.memberClusterList, ",")
+					aci.memberClusterList = ""
+					aci.parseClusterListFromKubeFedClusterList(clusterList, acm)
 					break
 				}
 			}
@@ -96,6 +70,26 @@ func (aci *AggregationClusterInfo) SetClusterInfo(acm *AggregationConfigMapInfo)
 	klog.Infoln("Get memberClusterList: " + aci.memberClusterList)
 }
 
+
+// parseClusterListFromKubeFedClusterList function is to setting the aci.memberClusterList
+// from kubeFedClusterList, which is dealt with the action of Upper and IgnorePrefix.
+func (aci *AggregationClusterInfo)parseClusterListFromKubeFedClusterList(clusterList *fedv1b1.
+	KubeFedClusterList, acm *AggregationConfigMapInfo) {
+	for _, cluster := range clusterList.Items {
+		var clusterTmp string
+		if acm.GetClusterIgnorePrefix() != "" {
+			clusterTmp = strings.TrimPrefix(cluster.Name,
+				acm.GetClusterIgnorePrefix())
+		} else {
+			clusterTmp = cluster.Name
+		}
+		aci.memberClusterList += strings.ToUpper(clusterTmp) + ","
+	}
+	aci.memberClusterList = strings.TrimRight(aci.memberClusterList, ",")
+}
+
+
+// GetClusterList return the memberClusterList to the public Callers
 func (aci *AggregationClusterInfo) GetClusterList() string {
 	return aci.memberClusterList
 }
